LayoutTests/imported/w3c/web-platform-tests/navigation-api/focus-reset/focus-reset-timing-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: undefined

Harness Error (FAIL), message = Unhandled rejection

PASS Focus should be reset before navigatesuccess
PASS Focus should be reset before navigateerror
